move GroupCaretakers
