+++ frontend/src/employee-frontend/App.tsx
@@ -20,7 +20,6 @@ import { featureFlags } from 'lib-customizations/employee'
 import EmployeeRoute from './components/EmployeeRoute'
 import FinancePage from './components/FinancePage'
 import { Footer } from './components/Footer'
-import GroupCaretakers from './components/GroupCaretakers'
 import Header from './components/Header'
 import IncomeStatementPage from './components/IncomeStatementPage'
 import LoginPage from './components/LoginPage'
@@ -118,6 +117,7 @@ import VoucherServiceProviders from './components/reports/VoucherServiceProvider
 import SettingsPage from './components/settings/SettingsPage'
 import SystemNotificationsPage from './components/system-notifications/SystemNotificationsPage'
 import TimelinePage from './components/timeline/TimelinePage'
+import GroupCaretakers from './components/unit/group-caretakers/GroupCaretakers'
 import DiscussionReservationSurveyWrapper from './components/unit/tab-calendar/discussion-surveys/DiscussionSurveyWrapper'
 import DiscussionReservationSurveysPage from './components/unit/tab-calendar/discussion-surveys/DiscussionSurveysPage'
 import CreateUnitPage from './components/unit/unit-details/CreateUnitPage'
